Detect 2 more statuses and support ESP-01 1M board (#5)
* [TASK] Add support for ESP-01 1M board

* [TASK] Properly detect status "Outside working area"

* [TASK] Properly detect status "Right wheel motor blocked"

* [TASK] Describe the mode when choosing to charge then back to AUTO

This is the mode that is published while the automower moves back
to the base station to charge and the resume mowing in AUTO mode.

* [TASK] Properly detect status "Waiting timer"

This is the code reported on a 230 ACX. We choose to add a
"2" suffix similarly to the 0x03F4 code, just in case, to
be able to differentiate between the two status codes.